requirements.txt
numpy
matplotlib
pyyaml
scipy
tabulate
opencv-python
tqdm
gnupg
pycrypto
rospkg